refactored the code
