+++ .gitignore
@@ -161,3 +161,6 @@ cython_debug/
 
 .DS_Store
 output
+
+data/
+figures/
