+++ README.md
@@ -1,0 +1,30 @@
+# Optimal Control for Minimum-Fuel Pinpoint Landing
+## AA 203: Optimal & Learning Based Control
+Devin Ardeshna, Nick Delurgio, and Pol Francesch Huc
+
+---
+
+## Introduction
+
+This repository encompasses the work the team conducted as part of AA 203 at Stanford Spring 2023. A full report can be found in this repository.
+
+## Environment Setup
+This environment setup assumes you have installed Python (version >= 3.8) on your machine. 
+```
+python -m venv aa203_final_project
+```
+Windows:
+```
+aa203_final_project\Scripts\activate
+```
+MacOS/Linux:
+```
+source aa203_final_project/bin/activate
+```
+
+Installing libaries:
+```
+pip install -r ./requirements.txt
+```
+
+## Running the code
