+++ requirements.txt
@@ -1,0 +1,8 @@
+control==0.9.3.post2
+cvxpy==1.3.1
+jax==0.4.8
+matplotlib==3.7.1
+numpy==1.24.2
+scipy==1.10.1
+sympy==1.12
+tqdm==4.65.0
